requirements.txt
aniso8601==9.0.1
attrs==21.4.0
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
dnspython==2.2.1
Flask==2.1.3
Flask-Bcrypt==1.0.1
flask-expects-json==1.7.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.7.2
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.4.0
pymongo==4.1.1
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.1
